Update CommunityResourcePack.netkan
While it looks like an updated release of CRP is currently in the pipeline, the only change seems to be with versioning.

This overrides the currently available (0.5.1.0) release of CRP to state that it is compatible with KSP up to 1.1.99.  Since this is only configs, it's exceedingly unlikely a minor version bump in KSP would break anything.  The CRP release in the pipeline has a corrected .version file and thus won't need the override going forward.